### 2018-11-06 #1067 @brainstorm
Fantastic! 👍

May I suggest a suitable reviewer? @reisingerf, would you like to review this tool for JOSS?
