### 2018-11-07 #1069 @whedon
[REVIEW]: hotsub - Command line tool to run batch jobs with ETL framework concurrently on cloud services
**Submitting author:** @otiai10 (<a href="http://orcid.org/0000-0001-6636-856X">Hiromu Ochiai</a>)
**Repository:** <a href="https://github.com/otiai10/hotsub" target ="_blank">https://github.com/otiai10/hotsub</a>
**Version:** v0.10.0
