### 2018-11-07 #1069 @brainstorm
@otiai10, could you please correct your `paper.bib`? See error above. Also, in the statement of need, I think that https://github.com/base2genomics/batchit should be pointed out perhaps?
