### 2018-11-08 #1069 @otiai10
@brainstorm
Thank you. I will

- [ ] Fix `bib` format
- [ ] Update `README.md` and `paper.md` to specify the statement of need more clearly
- [ ] Write available functionalities on `README.md` more clearly
- [ ] Add `CONTRIBUTING.md`
