### 2018-11-08 e7bab2e
List available commands and options in README.md
