### 2018-11-08 67f05a9
Fix description of command line to specify supports of GCP
