### 2018-11-08 3d2f506
Add CONTRIBUTING.md
