+++ CONTRIBUTING.md
@@ -1,0 +1,10 @@
+# Want to contribute?
+
+Any pull-request or posting issues are welcomed!
+
+- Check open issues for what you want.
+- If there is an open issue, comment on it. Otherwise open an issue describing your bug or feature with use cases.
+- Write code to fix the problem, then open a pull request with tests and documentation.
+    - Better add test cases to specify the problem under `test/cases`
+- The pull requests gets reviewed and then merged assuming there are no problems.
+- A new release version gets cut.
