### 2018-11-08 7024e43
Add "statement of need" in README.md
Related: openjournals/joss-reviews#1069
