### 2018-11-09 #1069 @reisingerf
@otiai10 could you please also put the output of `hotsub run --help` in the `README.md`?
It shows a lot of options and default values that are mentioned nowhere in the documentation.
