### 2018-11-11 b604969
Specify "statement of need" in paper.md
