### 2018-11-12 #1069 @otiai10
Thank you very much for taking time for review. I've patched some updates (mentioned above) to `bib` and `md`.
https://github.com/otiai10/hotsub
