### 2018-11-12 #1069 @brainstorm
@reisingerf @otiai10 Thanks much everyone!

@arfon LGTM, shall I deposit?
