### 2018-11-12 #1069 @arfon
@brainstorm - before we accept this, can we update the references to be actual citations in the text?

@otiai10 - Could you move the references you currently have in the `paper.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))

Also, it looks like we need to add a date to the `paper.md` file.
