### 2018-11-12 #1069 @otiai10
Thank you,

- I think all the references on `paper.md` are [listed on `paper.md`](https://github.com/otiai10/hotsub/tree/master/paper). Could you please specify what I should actually do more than this?
- Should I add `date` to the `paper.md` as the date it put on [`Zenodo`](https://zenodo.org/record/1482983#.W-myZXozbOR)? : November 12, 2018
