### 2018-11-12 #1069 @arfon
> - I think all the references on paper.md are listed on paper.md. Could you please specify what I should actually do more than this?

Yes, you need to remove all of the content underneath the [references section here](https://github.com/otiai10/hotsub/blob/master/paper/paper.md#references) and instead, cite the entries in your `paper.bib` inline in your paper body. When we compile the paper, the references will be compiled automatically.

> -  Should I add date to the paper.md as the date it put on Zenodo? : November 12, 2018

That should be fine yes.
