### 2018-11-13 d3a72bc
Remove contents of "Reference" section
so that automatically generated contents can be rendered underneath.
Related: openjournals/joss-reviews#1069
Link: https://github.com/openjournals/joss-reviews/issues/1069#issuecomment-437962573
