### 2018-11-13 dcf3ec4
Add "date" of publishing
as it got put on Zenodo
Related: openjournals/joss-reviews#1069
