+++ paper/paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
     - name: National Cancer Center Japan
       index: 1
-date: TBD
+date: November 12, 2018
 bibliography: paper.bib
 ---
 
