### 2018-11-13 #1069 @arfon
OK, final step now @otiai10 is to actually cite the references in your paper text. For example, [this citation](https://github.com/otiai10/hotsub/blob/master/paper/paper.bib#L9-L14) for Azure needs to be something like:

```
Azure Batch by Microsoft [@AzureBatch] ...
```

This `[@AzureBatch]` syntax is the bibtex entry key and will be used by Pandoc to compile the references section. This section is now currently blank as you're not actually citing any of the bibtex entries in the body of your paper.
