### 2018-11-13 415e422
Link references from bib to paper.md with Pandoc syntax.
See https://pandoc.org/
Related: openjournals/joss-reviews#1069
