### 2018-11-13 #1069 @arfon
@reisingerf - many thanks for your review here and to @brainstorm for editing this submission.

@otiai10 - your paper is now accepted into JOSS! :zap: :rocket: :boom:
