### 2018-11-14 #1069 @otiai10
-------

@arfon
CC @friend1ws

Thank you for everything.

I'd like to ask if it's possible to add **co-authors** and **acknowledgements**, and update the paper.

[This change](https://github.com/otiai10/hotsub/pull/119/files) is already merged to master and ready to generate pdf.

Thank you.
