### 2018-11-14 #1069 @arfon
> I'd like to ask if it's possible to add co-authors and acknowledgements, and update the paper.

Sure thing. The paper is now updated.
